+++ src/common/common.h
@@ -46,7 +46,7 @@ struct flow_metadata
 unsigned int read_exact(int fd, char *buf, size_t count, size_t max_per_read, bool dummy_buf);
 
 /* write exactly 'count' bytes into a socket 'fd' */
-unsigned int write_exact(int fd, char *buf, size_t count, size_t max_per_write,
+unsigned long long write_exact(int fd, char *buf, size_t count, size_t max_per_write,
     unsigned int rate_mbps, unsigned int tos, unsigned int sleep_overhead_us, bool dummy_buf);
 
 /* read the metadata of a flow from a socket and return true if it succeeds. */
